Moved AdvancedPanel to top level
