+++ src/main/java/me/jsedwards/configserver/ServerConfigPanel.java
@@ -3,15 +3,11 @@ package me.jsedwards.configserver;
 import me.jsedwards.Card;
 import me.jsedwards.Main;
 import me.jsedwards.dashboard.Server;
-import me.jsedwards.util.UnifiedListenerTextField;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
 import javax.swing.*;
-import javax.swing.text.DefaultStyledDocument;
 import java.awt.*;
-import java.util.function.Function;
-import java.util.function.Predicate;
 
 public class ServerConfigPanel extends JPanel implements Card {
 
@@ -67,7 +63,7 @@ public class ServerConfigPanel extends JPanel implements Card {
     @Override
     public void exit() {
         for (AdvancedPanel panel : advancedPanels) {
-            panel.properties.save();
+            panel.saveProperties();
         }
     }
 
@@ -95,125 +91,11 @@ public class ServerConfigPanel extends JPanel implements Card {
             slider.addChangeListener(e -> {
                 int sliderValue = slider.getValue();
                 for (AdvancedPanel panel : ServerConfigPanel.this.advancedPanels) {
-                    panel.properties.optimise(sliderValue);
+                    panel.optimise(sliderValue);
                 }
             });
             return slider;
         }
     }
 
-    private static class AdvancedPanel extends JPanel {
-
-        private final SidePanel sidePanel;
-        private final Function<Server, ConfigManager> configManagerCreator;
-        private final Predicate<Server> enabled;
-        private ConfigManager properties;
-        private final JList<String> propertiesList;
-        public final String name;
-
-        public AdvancedPanel(Function<Server, ConfigManager> configManagerCreator, String name, Predicate<Server> enabled) {
-            this.name = name;
-            this.configManagerCreator = configManagerCreator;
-            this.enabled = enabled;
-            this.setLayout(new GridBagLayout());
-            properties = configManagerCreator.apply(null);
-            // Search label
-            this.add(new JLabel("Search properties:"), new GridBagConstraints(1, 1, 1, 1, 0, 0, GridBagConstraints.WEST, GridBagConstraints.NONE, new Insets(0, 0, 0, 0), 0, 0));
-            // Search box
-            JTextField searchBox = new UnifiedListenerTextField() {
-                @Override
-                protected void update() {
-                    properties.updateSearch(this.getText());
-                    AdvancedPanel.this.updateList();
-                }
-            };
-            this.add(searchBox, new GridBagConstraints(2, 1, 1, 1, 1, 0, GridBagConstraints.EAST, GridBagConstraints.HORIZONTAL, new Insets(0, 0, 0, 0), 0, 0));
-            // List box
-            propertiesList = new JList<>(properties);
-            JScrollPane scrollPane = new JScrollPane(propertiesList, JScrollPane.VERTICAL_SCROLLBAR_AS_NEEDED, JScrollPane.HORIZONTAL_SCROLLBAR_AS_NEEDED);
-            this.add(scrollPane, new GridBagConstraints(1, 2, 2, 1, 1, 1, GridBagConstraints.NORTH, GridBagConstraints.BOTH, new Insets(0, 0, 0, 0), 0, 0));
-            // Side panel
-            sidePanel = new SidePanel();
-            this.add(sidePanel, new GridBagConstraints(3, 1, 1, 2, 0.3, 1, GridBagConstraints.NORTH, GridBagConstraints.BOTH, new Insets(0, 0, 0, 0), 0, 0));
-            // Selection listener
-            propertiesList.addListSelectionListener(e -> sidePanel.update());
-        }
-
-        private boolean isEnabled(Server server) {
-            return enabled.test(server);
-        }
-
-        private void setServer(Server server) {
-            this.properties = configManagerCreator.apply(server);
-            this.propertiesList.setModel(this.properties);
-        }
-
-        private void updateList() {
-            propertiesList.invalidate();
-            propertiesList.repaint();
-            sidePanel.update();
-        }
-
-        private class SidePanel extends JPanel {
-
-            private final JLabel nameLbl;
-            private final JLabel dataTypeLbl;
-            private final JLabel defaultValueLbl;
-            private final JTextPane descriptionLbl;
-
-            public SidePanel() {
-                super();
-                setLayout(new GridBagLayout());
-                // Property name
-                nameLbl = new JLabel("Select a property");
-                nameLbl.setFont(Main.MAIN_FONT.deriveFont(18f));
-                this.add(nameLbl, new GridBagConstraints(1, 1, 1, 1, 1, 0, GridBagConstraints.NORTH, GridBagConstraints.HORIZONTAL, new Insets(0, 5, 0, 0), 0, 0));
-                // Data type
-                dataTypeLbl = new JLabel("Data type:");
-                this.add(dataTypeLbl, new GridBagConstraints(1, 2, 1, 1, 1, 0, GridBagConstraints.NORTH, GridBagConstraints.HORIZONTAL, new Insets(10, 5, 0, 0), 0, 0));
-                // Default value
-                defaultValueLbl = new JLabel("Default value:");
-                this.add(defaultValueLbl, new GridBagConstraints(1, 3, 1, 1, 1, 0, GridBagConstraints.NORTH, GridBagConstraints.HORIZONTAL, new Insets(10, 5, 0, 0), 0, 0));
-                // Property description
-                descriptionLbl = new JTextPane(new DefaultStyledDocument());
-                descriptionLbl.setContentType("text/html");
-                descriptionLbl.setEditable(false);
-                this.add(descriptionLbl, new GridBagConstraints(1, 4, 1, 1, 1, 0, GridBagConstraints.NORTH, GridBagConstraints.HORIZONTAL, new Insets(20, 0, 0, 0), 0, 0));
-                // Edit button
-                JButton editBtn = createEditButton();
-                this.add(editBtn, new GridBagConstraints(1, 5, 1, 1, 1, 0, GridBagConstraints.NORTH, GridBagConstraints.HORIZONTAL, new Insets(10, 5, 0, 0), 0, 0));
-                // Padding
-                this.add(new JPanel(), new GridBagConstraints(1, 6, 1, 1, 1, 1, GridBagConstraints.NORTH, GridBagConstraints.BOTH, new Insets(0, 0, 0, 0), 0, 0));
-            }
-
-            @Override
-            public Dimension getPreferredSize() {
-                return new Dimension((int) (AdvancedPanel.this.getWidth() * 0.05), AdvancedPanel.this.getHeight());
-            }
-
-            private JButton createEditButton() {
-                JButton editBtn = new JButton("Edit");
-                editBtn.addActionListener(e -> {
-                    String selected = AdvancedPanel.this.propertiesList.getSelectedValue();
-                    int splitIndex = selected.indexOf(':');
-                    String key = selected.substring(0, splitIndex);
-                    String value = JOptionPane.showInputDialog(editBtn, "Enter new value for %s:".formatted(key), "Edit value", JOptionPane.QUESTION_MESSAGE);
-                    AdvancedPanel.this.properties.set(key, value);
-                    AdvancedPanel.this.updateList();
-                });
-                return editBtn;
-            }
-
-            private void update() {
-                String selectedItem = AdvancedPanel.this.propertiesList.getSelectedValue();
-                if (selectedItem != null) {
-                    String key = selectedItem.substring(0, selectedItem.indexOf(':'));
-                    this.nameLbl.setText(key);
-                    this.descriptionLbl.setText("<html>" + AdvancedPanel.this.properties.getDescription(key) + "</html>");
-                    this.dataTypeLbl.setText("Data type: " + AdvancedPanel.this.properties.getDataType(key));
-                    this.defaultValueLbl.setText("Default value: " + AdvancedPanel.this.properties.getDefaultValue(key));
-                }
-            }
-        }
-    }
 }
